global gui state working i think :)

+++ SBBB_Application/include/util/Array3D.hpp
@@ -38,7 +38,7 @@ public:
 
 	T& operator()(size_t x, size_t y, size_t z) {
 #ifdef SBBB_DEBUG
-		if (!bounded(x, y, z)) {
+		if (!bounded((int)x, (int)y, (int)z)) {
 			throw std::out_of_range("3D Array index out of bounds.");
 		}
 #endif
@@ -51,7 +51,7 @@ public:
 	}
 	T& operator()(int x, int y, int z) {
 #ifdef SBBB_DEBUG
-		if (!bounded((size_t)x, (size_t)y, (size_t)z)) {
+		if (!bounded(x, y, z)) {
 			throw std::out_of_range("3D Array index out of bounds.");
 		}
 #endif
